Move to dotnet 8 Nullable reference types

+++ src/Data/SparkApplicationService.cs
@@ -10,6 +10,7 @@ using Microsoft.Extensions.Caching.Memory;
 using k8s;
 
 using KubeStatus.Models;
+using Microsoft.VisualBasic;
 
 namespace KubeStatus.Data
 {
@@ -21,56 +22,63 @@ namespace KubeStatus.Data
 
         private static readonly JsonSerializerOptions _jsonSerializerOptions = new() { PropertyNamingPolicy = JsonNamingPolicy.CamelCase };
 
-        public async Task<IEnumerable<SparkApplication>> GetAllSparkApplicationsAsync()
+        public async Task<IEnumerable<SparkApplication>?> GetAllSparkApplicationsAsync()
         {
             return await GetSparkApplicationsAsync();
         }
 
-        public Task<IEnumerable<SparkApplication>> GetSparkApplicationsAsync(string filterStatus = null)
+        public Task<IEnumerable<SparkApplication>?> GetSparkApplicationsAsync(string? filterStatus = null)
         {
-            return MemoryCache.GetOrCreateAsync(filterStatus, async e =>
+            var sparkApplications = new List<SparkApplication>();
+
+            if (!string.IsNullOrWhiteSpace(filterStatus))
             {
-                e.SetOptions(new MemoryCacheEntryOptions
+                return MemoryCache.GetOrCreateAsync(filterStatus, async e =>
                 {
-                    AbsoluteExpirationRelativeToNow =
-                        TimeSpan.FromSeconds(10)
-                });
-
-                var sparkApplications = new List<SparkApplication>();
-
-                var response = await kubernetesClient.CustomObjects.ListClusterCustomObjectAsync(Helper.SparkGroup(), Helper.SparkApplicationVersion(), Helper.SparkApplicationPlural());
-
-                var jsonString = JsonSerializer.Serialize<object>(response);
-                JsonNode jsonNode = JsonNode.Parse(jsonString)!;
-                JsonNode itemsNode = jsonNode!["items"]!;
-
-                foreach (var item in itemsNode.AsArray())
-                {
-                    var status = JsonSerializer.Deserialize<SparkApplicationStatus>(item!["status"]!, _jsonSerializerOptions);
-                    var applicationState = status?.ApplicationState["state"];
-
-                    if (!string.IsNullOrWhiteSpace(filterStatus) && !filterStatus.Equals(applicationState, StringComparison.OrdinalIgnoreCase))
+                    e.SetOptions(new MemoryCacheEntryOptions
                     {
-                        continue;
+                        AbsoluteExpirationRelativeToNow =
+                            TimeSpan.FromSeconds(10)
+                    });
+
+                    var sparkApplications = new List<SparkApplication>();
+
+                    var response = await kubernetesClient.CustomObjects.ListClusterCustomObjectAsync(Helper.SparkGroup(), Helper.SparkApplicationVersion(), Helper.SparkApplicationPlural());
+
+                    var jsonString = JsonSerializer.Serialize<object>(response);
+                    JsonNode jsonNode = JsonNode.Parse(jsonString)!;
+                    JsonNode itemsNode = jsonNode!["items"]!;
+
+                    foreach (var item in itemsNode.AsArray())
+                    {
+                        var status = JsonSerializer.Deserialize<SparkApplicationStatus>(item!["status"]!, _jsonSerializerOptions);
+                        var applicationState = status?.ApplicationState?["state"] ?? "Unknown";
+
+                        if (!filterStatus.Equals(applicationState, StringComparison.OrdinalIgnoreCase))
+                        {
+                            continue;
+                        }
+
+                        var apiVersion = item!["apiVersion"]!.ToString();
+                        var kind = item!["kind"]!.ToString();
+                        var metadata = JsonSerializer.Deserialize<Metadata>(item!["metadata"]!, _jsonSerializerOptions);
+                        var spec = item!["spec"]!.ToString();
+
+                        sparkApplications.Add(new SparkApplication
+                        {
+                            ApiVersion = apiVersion,
+                            Kind = kind,
+                            Metadata = metadata,
+                            Status = status,
+                            Spec = spec
+                        });
                     }
 
-                    var apiVersion = item!["apiVersion"]!.ToString();
-                    var kind = item!["kind"]!.ToString();
-                    var metadata = JsonSerializer.Deserialize<Metadata>(item!["metadata"]!, _jsonSerializerOptions);
-                    var spec = item!["spec"]!.ToString();
+                    return sparkApplications.AsEnumerable();
+                });
+            }
 
-                    sparkApplications.Add(new SparkApplication
-                    {
-                        ApiVersion = apiVersion,
-                        Kind = kind,
-                        Metadata = metadata,
-                        Status = status,
-                        Spec = spec
-                    });
-                }
-
-                return sparkApplications.AsEnumerable();
-            });
+            return (Task<IEnumerable<SparkApplication>?>)sparkApplications.AsEnumerable();
         }
 
         public async Task<int> DeleteFailedSparkApplicationsAsync(string keepHours = "168")
@@ -81,23 +89,29 @@ namespace KubeStatus.Data
                 {
                     hours = 168;
                 }
-                hours = hours * -1;
+                hours *= -1;
                 var date = DateTime.UtcNow.AddHours(hours);
 
                 var crs = await GetSparkApplicationsAsync("failed");
                 var deleted = 0;
-                foreach (var cr in crs)
+                if (crs != null)
                 {
-                    if (cr.Metadata.CreationTimestamp.UtcDateTime < date)
+                    foreach (var cr in crs)
                     {
-                        await kubernetesClient.CustomObjects.DeleteNamespacedCustomObjectAsync(
-                            Helper.SparkGroup(),
-                            Helper.SparkApplicationVersion(),
-                            cr.Metadata.Namespace ?? "default",
-                            Helper.SparkApplicationPlural(),
-                            cr.Metadata.Name).ConfigureAwait(false);
+                        if (cr != null && cr.Metadata != null)
+                        {
+                            if (cr.Metadata.CreationTimestamp.UtcDateTime < date)
+                            {
+                                await kubernetesClient.CustomObjects.DeleteNamespacedCustomObjectAsync(
+                                    Helper.SparkGroup(),
+                                    Helper.SparkApplicationVersion(),
+                                    cr.Metadata.Namespace ?? "default",
+                                    Helper.SparkApplicationPlural(),
+                                    cr.Metadata.Name).ConfigureAwait(false);
 
-                        deleted++;
+                                deleted++;
+                            }
+                        }
                     }
                 }
                 return deleted;
